+++ docker/docker-compose.yml
@@ -8,7 +8,7 @@ services:
     volumes:
       - pg-data:/var/lib/postgresql
     ports:
-      - 5432:5432
+      - 5433:5432
 
 
 volumes:
