+++ src/main/java/com/summitsync/api/eventperiod/EventPeriod.java
@@ -1,4 +1,4 @@
-package com.summitsync.api.model;
+package com.summitsync.api.eventperiod;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
