--- src/main/java/com/summitsync/api/model/Trainer.java
@@ -1,22 +1,0 @@
-package com.summitsync.api.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Table(name = "SS_Trainer")
-public class Trainer {
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private long trainerId;
-    private String subjectId;
-    @ManyToMany(fetch = FetchType.LAZY)
-    private List<Qualification> qualifications;
-}
