+++ src/main/java/com/summitsync/api/qualification/Qualification.java
@@ -1,5 +1,8 @@
-package com.summitsync.api.model;
+package com.summitsync.api.qualification;
 
+import com.summitsync.api.course.Course;
+import com.summitsync.api.group.Group;
+import com.summitsync.api.trainer.Trainer;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
