+++ src/main/resources/application.properties
@@ -1,8 +1,7 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/summitsync
+spring.datasource.url=jdbc:postgresql://localhost:5433/summitsync
 spring.datasource.username=dev
 spring.datasource.password=dev
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 logging.level.web=debug
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.security.oauth2.resourceserver.jwt.issuer-uri=https://keycloak.summitsync.meschter.me/realms/summit-sync
